Se arregla funcion consultar y se prueba en postman, faltan arreglos para el postmapping y solucionar error en el controller que muestra mensaje de cliente creado exitosamente pero no envia nada al servidor para guardar en la base de datos

+++ src/main/java/co/com/cmdb/entity/ClienteEntity.java
@@ -47,7 +47,7 @@ public final class ClienteEntity {
 		return new ClienteEntity();
 	}
 	
-	public static final ClienteEntity build( final String identificador, final TipoDocumentoEntity tipoDocumento, final String nombre, final String apellidoss, final String correo, final int telefono, final boolean estado) {
+	public static final ClienteEntity build( final String identificador, final TipoDocumentoEntity tipoDocumento, final String nombre, final String apellidoss, final String correo, final long telefono, final boolean estado) {
 		return new ClienteEntity( identificador, tipoDocumento, nombre, apellidoss,correo,telefono, estado);
 	}
